backend/requirements.txt
python-dotenv
pydantic
requests
openapi-spec-validator
starlette
connexion[async,swagger-ui,uvicorn]
cryptography
setuptools
packaging
alembic
sqlalchemy
sqlmodel
aiosqlite
asyncio
aiohttp
aioftp
aiofiles
structlog
greenlet
langchain_community
langchain_chroma
langchain_openai
pypdf
httpx
langchain_ollama
google-api-core==2.10.2
google-cloud-core==2.3.2
google-auth==2.12.0
google-cloud-documentai==2.0.0
googleapis-common-protos==1.56.4